+++ python/pynopegl/_pynopegl.pyx
@@ -268,6 +268,7 @@ cdef extern from "nopegl/nopegl.h":
                          const double *offsets, double v, double *t)
 
     int ngl_timerangefilter_set_range(ngl_node *node, double start, double end)
+    int ngl_timerangefilter2d_set_range(ngl_node *node, double start, double end)
 
 cdef extern from "nopegl/nopegl_opengl.h":
     cdef struct ngl_config_gl:
@@ -647,6 +648,9 @@ cdef class _Node:
 
     def _timerangefilter_set_range(self, double start, double end):
         return ngl_timerangefilter_set_range(self.ctx, start, end)
+
+    def _timerangefilter2d_set_range(self, double start, double end):
+        return ngl_timerangefilter2d_set_range(self.ctx, start, end)
 
 ANIM_EVALUATE, ANIM_DERIVATE, ANIM_SOLVE = range(3)